requirements.txt
discord.py==2.1.0
jishaku==2.5.1
pydantic==1.10.4
python-dotenv==0.21.1
watchfiles==0.18.1
